extract components, allow to map formula for non topic related formulas
